# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
                required:
                  - categories
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryInput"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all payment sources/accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
                required:
                  - accounts
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountInput"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /email-import/parse:
    post:
      operationId: parseEmailTransaction
      tags: [email-import]
      summary: Parse email content and extract transaction data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailParseInput"
      responses:
        "200":
          description: Parsed transaction data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailParseResult"
  /email-import/import:
    post:
      operationId: importEmailTransactions
      tags: [email-import]
      summary: Import parsed transactions from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailImportResult"
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        categoryId:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        categoryColor:
          type: string
          nullable: true
        categoryIcon:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        accountName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        importSource:
          type: string
          enum: [manual, email]
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - amount
        - type
        - description
        - date
        - createdAt
        - updatedAt
    CreateTransactionInput:
      type: object
      properties:
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        categoryId:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        importSource:
          type: string
          enum: [manual, email]
          nullable: true
      required:
        - amount
        - type
        - description
        - date
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - icon
        - color
        - type
        - isDefault
        - createdAt
    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
      required:
        - name
        - icon
        - color
        - type
    UpdateCategoryInput:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [bank, credit_card, cash, wallet, investment, other]
        color:
          type: string
        icon:
          type: string
        openingBalance:
          type: number
          format: double
        currentBalance:
          type: number
          format: double
        currency:
          type: string
          default: INR
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - color
        - icon
        - openingBalance
        - currentBalance
        - currency
        - isDefault
        - createdAt
        - updatedAt
    CreateAccountInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [bank, credit_card, cash, wallet, investment, other]
        color:
          type: string
        icon:
          type: string
        openingBalance:
          type: number
          format: double
        currency:
          type: string
          default: INR
        isDefault:
          type: boolean
      required:
        - name
        - type
        - color
        - icon
        - openingBalance
    UpdateAccountInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [bank, credit_card, cash, wallet, investment, other]
        color:
          type: string
        icon:
          type: string
        openingBalance:
          type: number
          format: double
        currency:
          type: string
        isDefault:
          type: boolean
...
    EmailParseInput:
      type: object
      properties:
        emailContent:
          type: string
          description: Raw email body text or HTML
        emailSubject:
          type: string
          nullable: true
        emailFrom:
          type: string
          nullable: true
      required:
        - emailContent
    ParsedTransaction:
      type: object
      properties:
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        suggestedCategoryId:
          type: string
          nullable: true
        suggestedCategoryName:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        confidence:
          type: number
          format: double
          description: "Parsing confidence 0-1"
        rawText:
          type: string
          nullable: true
      required:
        - amount
        - type
        - description
        - date
        - confidence
    EmailParseResult:
      type: object
      properties:
        success:
          type: boolean
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/ParsedTransaction"
        rawExtracted:
          type: string
          nullable: true
        errors:
          type: array
          items:
            type: string
      required:
        - success
        - transactions
        - errors
    EmailImportInput:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/CreateTransactionInput"
      required:
        - transactions
    EmailImportResult:
      type: object
      properties:
        imported:
          type: integer
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - imported
        - transactions
...