# lib/db/src/schema/accounts.ts
import { pgTable, text, boolean, timestamp, uuid, doublePrecision, varchar } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { usersTable } from "./auth";

export const accountsTable = pgTable("accounts", {
  id: uuid("id").primaryKey().defaultRandom(),
  userId: varchar("user_id").references(() => usersTable.id, { onDelete: "cascade" }),
  name: text("name").notNull(),
  type: text("type", {
    enum: ["bank", "credit_card", "cash", "wallet", "investment", "other"],
  }).notNull().default("bank"),
  color: text("color").notNull().default("#6366f1"),
  icon: text("icon").notNull().default("🏦"),
  openingBalance: doublePrecision("opening_balance").notNull().default(0),
  currency: text("currency").notNull().default("INR"),
  isDefault: boolean("is_default").notNull().default(false),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertAccountSchema = createInsertSchema(accountsTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
});

export type InsertAccount = z.infer<typeof insertAccountSchema>;
export type Account = typeof accountsTable.$inferSelect;